fix(single-threaded): Use uint64_t for random seed
- Changed seed type from unsigned int to uint64_t
- Updated random number generation function call

+++ src/single-threaded.c
@@ -5,8 +5,10 @@
 
 double montecarlo(int64_t n) {
   int64_t count = 0;
+
   float x, y;
-  unsigned int seed = time(NULL);
+
+  uint64_t seed = time(NULL);
 
   for (int64_t i = 0; i < n; i++) {
     x = randomf(&seed);
